pyproject.toml
[project]
name = "data-talk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckdb>=1.1.3",
    "duckdb-engine>=0.15.0",
    "huggingface-hub>=0.27.1",
    "ipykernel>=6.29.5",
    "jupyter>=1.1.1",
    "langchain>=0.3.14",
    "langchain-community>=0.3.14",
    "langchain-google-genai>=2.0.9",
    "langgraph>=0.2.69",
    "llama-index>=0.12.11",
    "llama-index-llms-huggingface>=0.4.2",
    "pydantic-ai>=0.0.19",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.37",
]
